demand-forecasting/fastapi/requirements.txt
numpy
pandas
joblib
uvicorn
lightgbm
fastapi[all]
python-dateutil